requirements.txt
asttokens==2.0.5
backcall==0.2.0   
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==5.1.1
executing==0.8.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
peewee==3.15.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.12.0
pyTelegramBotAPI==4.6.0
requests==2.28.1
six==1.16.0
stack-data==0.3.0
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
telebot~=0.0.5
gspread~=5.11.3
oauth2client~=4.1.3
google-api-python-client~=2.107.0
DateTime~=4.4